fix: enhance leave room functionality
